backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
supabase==2.9.1
openai==1.51.0
PyJWT==2.9.0
python-multipart==0.0.12
python-dotenv==1.0.1
